Initial implenemtation of Core YAML parser
- Created new YAML parser
- Extracted common parts for YAML and MD to the separate classes
- Added unit tests
