+++ data/test_yaml_stories/stories_1.yml
@@ -1,0 +1,6 @@
+stories:
+- story: simple_story_without_checkpoint
+  steps:
+  - user: /simple
+  - action: utter_default
+  - action: utter_greet
