+++ data/test_yaml_stories/stories_2.yml
@@ -1,0 +1,7 @@
+stories:
+- story: simple_story_with_only_end
+  steps:
+  - user: /hello
+  - action: utter_greet
+  - slot: name
+    value: peter
