+++ rasa/core/training/story_reader/markdown_story_reader.py
@@ -1,0 +1,237 @@
+import asyncio
+import json
+import logging
+import os
+import re
+from typing import Optional, Dict, Text, List, Any
+
+from rasa.constants import DOCS_URL_DOMAINS, DOCS_URL_STORIES
+from rasa.core.constants import INTENT_MESSAGE_PREFIX
+from rasa.core.domain import Domain
+from rasa.core.events import UserUttered
+from rasa.core.exceptions import StoryParseError
+from rasa.core.interpreter import NaturalLanguageInterpreter, RegexInterpreter
+from rasa.core.training.dsl import EndToEndReader
+from rasa.core.training.story_reader.story_reader import StoryReader
+import rasa.utils.io as io_utils
+
+from rasa.core.training.structures import (
+    StoryStep,
+    FORM_PREFIX,
+)
+from rasa.utils.common import raise_warning
+
+logger = logging.getLogger(__name__)
+
+
+class MarkdownStoryReader(StoryReader):
+    def __init__(
+        self,
+        interpreter: NaturalLanguageInterpreter,
+        domain: Optional[Domain] = None,
+        template_vars: Optional[Dict] = None,
+        use_e2e: bool = False,
+        source_name: Text = None,
+    ):
+        super().__init__(interpreter, domain, template_vars, use_e2e, source_name)
+
+    async def read_from_file(self, filename: Text,) -> List[StoryStep]:
+        """Given a md file reads the contained stories."""
+
+        try:
+            with open(filename, "r", encoding=io_utils.DEFAULT_ENCODING) as f:
+                lines = f.readlines()
+
+            return await self._process_lines(lines)
+        except ValueError as err:
+            file_info = "Invalid story file format. Failed to parse '{}'".format(
+                os.path.abspath(filename)
+            )
+            logger.exception(file_info)
+            if not err.args:
+                err.args = ("",)
+            err.args = err.args + (file_info,)
+            raise
+
+    async def _process_lines(self, lines: List[Text]) -> List[StoryStep]:
+        multiline_comment = False
+
+        for idx, line in enumerate(lines):
+            line_num = idx + 1
+            try:
+                line = self._replace_template_variables(self._clean_up_line(line))
+                if line.strip() == "":
+                    continue
+                elif line.startswith("<!--"):
+                    multiline_comment = True
+                    continue
+                elif multiline_comment and line.endswith("-->"):
+                    multiline_comment = False
+                    continue
+                elif multiline_comment:
+                    continue
+                elif line.startswith("#"):
+                    # reached a new story block
+                    name = line[1:].strip("# ")
+                    self._new_story_part(name, self.source_name)
+                elif line.startswith(">"):
+                    # reached a checkpoint
+                    name, conditions = self._parse_event_line(line[1:].strip())
+                    self.add_checkpoint(name, conditions)
+                elif re.match(fr"^[*\-]\s+{FORM_PREFIX}", line):
+                    logger.debug(
+                        "Skipping line {}, "
+                        "because it was generated by "
+                        "form action".format(line)
+                    )
+                elif line.startswith("-"):
+                    # reached a slot, event, or executed action
+                    event_name, parameters = self._parse_event_line(line[1:])
+                    self._add_event(event_name, parameters)
+                elif line.startswith("*"):
+                    # reached a user message
+                    user_messages = [el.strip() for el in line[1:].split(" OR ")]
+                    if self.use_e2e:
+                        await self._add_e2e_messages(user_messages, line_num)
+                    else:
+                        await self._add_user_messages(user_messages, line_num)
+                else:
+                    # reached an unknown type of line
+                    logger.warning(
+                        f"Skipping line {line_num}. "
+                        "No valid command found. "
+                        f"Line Content: '{line}'"
+                    )
+            except Exception as e:
+                msg = f"Error in line {line_num}: {e}"
+                logger.error(msg, exc_info=1)  # pytype: disable=wrong-arg-types
+                raise ValueError(msg)
+        self._add_current_stories_to_result()
+        return self.story_steps
+
+    @staticmethod
+    def _parameters_from_json_string(s: Text, line: Text) -> Dict[Text, Any]:
+        """Parse the passed string as json and create a parameter dict."""
+
+        if s is None or not s.strip():
+            # if there is no strings there are not going to be any parameters
+            return {}
+
+        try:
+            parsed_slots = json.loads(s)
+            if isinstance(parsed_slots, dict):
+                return parsed_slots
+            else:
+                raise Exception(
+                    "Parsed value isn't a json object "
+                    "(instead parser found '{}')"
+                    ".".format(type(parsed_slots))
+                )
+        except Exception as e:
+            raise ValueError(
+                "Invalid to parse arguments in line "
+                "'{}'. Failed to decode parameters"
+                "as a json object. Make sure the event"
+                "name is followed by a proper json "
+                "object. Error: {}".format(line, e)
+            )
+
+    def _replace_template_variables(self, line: Text) -> Text:
+        def process_match(matchobject):
+            varname = matchobject.group(1)
+            if varname in self.template_variables:
+                return self.template_variables[varname]
+            else:
+                raise ValueError(
+                    "Unknown variable `{var}` "
+                    "in template line '{line}'"
+                    "".format(var=varname, line=line)
+                )
+
+        template_rx = re.compile(r"`([^`]+)`")
+        return template_rx.sub(process_match, line)
+
+    @staticmethod
+    def _clean_up_line(line: Text) -> Text:
+        """Removes comments and trailing spaces"""
+
+        return re.sub(r"<!--.*?-->", "", line).strip()
+
+    def add_checkpoint(self, name: Text, conditions: Optional[Dict[Text, Any]]) -> None:
+
+        # Ensure story part already has a name
+        if not self.current_step_builder:
+            raise StoryParseError(
+                "Checkpoint '{}' is at an invalid location. "
+                "Expected a story start.".format(name)
+            )
+
+        self.current_step_builder.add_checkpoint(name, conditions)
+
+    @staticmethod
+    def _parse_event_line(line):
+        """Tries to parse a single line as an event with arguments."""
+
+        # the regex matches "slot{"a": 1}"
+        m = re.search("^([^{]+)([{].+)?", line)
+        if m is not None:
+            event_name = m.group(1).strip()
+            slots_str = m.group(2)
+            parameters = MarkdownStoryReader._parameters_from_json_string(
+                slots_str, line
+            )
+            return event_name, parameters
+        else:
+            raise_warning(
+                f"Failed to parse action line '{line}'. Ignoring this line.",
+                docs=DOCS_URL_STORIES,
+            )
+            return "", {}
+
+    async def _add_user_messages(self, messages, line_num):
+        if not self.current_step_builder:
+            raise StoryParseError(
+                "User message '{}' at invalid location. "
+                "Expected story start.".format(messages)
+            )
+        parsed_messages = await asyncio.gather(
+            *[self._parse_message(m, line_num) for m in messages]
+        )
+        self.current_step_builder.add_user_messages(parsed_messages)
+
+    async def _add_e2e_messages(self, e2e_messages: List[Text], line_num: int) -> None:
+        if not self.current_step_builder:
+            raise StoryParseError(
+                "End-to-end message '{}' at invalid "
+                "location. Expected story start."
+                "".format(e2e_messages)
+            )
+        e2e_reader = EndToEndReader()
+        parsed_messages = []
+        for m in e2e_messages:
+            message = e2e_reader._parse_item(m)
+            parsed = await self._parse_message(message.text, line_num)
+
+            parsed.parse_data["true_intent"] = message.data["true_intent"]
+            parsed.parse_data["true_entities"] = message.data.get("entities") or []
+            parsed_messages.append(parsed)
+        self.current_step_builder.add_user_messages(parsed_messages)
+
+    async def _parse_message(self, message: Text, line_num: int) -> UserUttered:
+        if message.startswith(INTENT_MESSAGE_PREFIX):
+            parse_data = await RegexInterpreter().parse(message)
+        else:
+            parse_data = await self.interpreter.parse(message)
+        utterance = UserUttered(
+            message, parse_data.get("intent"), parse_data.get("entities"), parse_data
+        )
+        intent_name = utterance.intent.get("name")
+        if self.domain and intent_name not in self.domain.intents:
+            raise_warning(
+                f"Found unknown intent '{intent_name}' on line {line_num}. "
+                "Please, make sure that all intents are "
+                "listed in your domain yaml.",
+                UserWarning,
+                docs=DOCS_URL_DOMAINS,
+            )
+        return utterance
