+++ rasa/core/training/story_reader/story_reader.py
@@ -1,0 +1,73 @@
+import asyncio
+import logging
+from typing import Optional, Dict, Text, List
+
+from rasa.constants import DOCS_URL_DOMAINS
+from rasa.core.constants import INTENT_MESSAGE_PREFIX
+from rasa.core.domain import Domain
+from rasa.core.events import SlotSet, ActionExecuted, Event, UserUttered
+from rasa.core.exceptions import StoryParseError
+from rasa.core.interpreter import NaturalLanguageInterpreter, RegexInterpreter
+from rasa.core.training.dsl import EndToEndReader
+from rasa.core.training.story_reader.story_step_builder import StoryStepBuilder
+from rasa.core.training.structures import StoryStep
+from rasa.utils.common import raise_warning
+
+logger = logging.getLogger(__name__)
+
+
+class StoryReader:
+    """Helper class to read a story file."""
+
+    def __init__(
+        self,
+        interpreter: NaturalLanguageInterpreter,
+        domain: Optional[Domain] = None,
+        template_vars: Optional[Dict] = None,
+        use_e2e: bool = False,
+        source_name: Text = None,
+    ):
+        self.story_steps = []
+        self.current_step_builder: Optional[StoryStepBuilder] = None
+        self.domain = domain
+        self.interpreter = interpreter
+        self.template_variables = template_vars if template_vars else {}
+        self.use_e2e = use_e2e
+        self.source_name = source_name
+
+    async def read_from_file(self, filename: Text,) -> List[StoryStep]:
+        raise NotImplementedError
+
+    def _add_current_stories_to_result(self):
+        if self.current_step_builder:
+            self.current_step_builder.flush()
+            self.story_steps.extend(self.current_step_builder.story_steps)
+
+    def _new_story_part(self, name: Text, source_name: Text):
+        self._add_current_stories_to_result()
+        self.current_step_builder = StoryStepBuilder(name, source_name)
+
+    def _add_event(self, event_name, parameters):
+
+        # add 'name' only if event is not a SlotSet,
+        # because there might be a slot with slot_key='name'
+        if "name" not in parameters and event_name != SlotSet.type_name:
+            parameters["name"] = event_name
+
+        parsed_events = Event.from_story_string(
+            event_name, parameters, default=ActionExecuted
+        )
+        if parsed_events is None:
+            raise StoryParseError(
+                "Unknown event '{}'. It is Neither an event "
+                "nor an action).".format(event_name)
+            )
+        if self.current_step_builder is None:
+            raise StoryParseError(
+                "Failed to handle event '{}'. There is no "
+                "started story block available. "
+                "".format(event_name)
+            )
+
+        for p in parsed_events:
+            self.current_step_builder.add_event(p)
