+++ rasa/importers/rasa.py
@@ -5,7 +5,6 @@ from typing import Dict, List, Optional, Text, Union
 from rasa import data
 from rasa.core.domain import Domain, InvalidDomain
 from rasa.core.interpreter import NaturalLanguageInterpreter, RegexInterpreter
-from rasa.core.training.dsl import StoryFileReader
 from rasa.core.training.structures import StoryGraph
 from rasa.importers import utils
 from rasa.importers.importer import TrainingDataImporter
@@ -47,7 +46,7 @@ class RasaFileImporter(TrainingDataImporter):
         exclusion_percentage: Optional[int] = None,
     ) -> StoryGraph:
 
-        story_steps = await StoryFileReader.read_from_files(
+        return await utils.story_graph_from_paths(
             self._story_files,
             await self.get_domain(),
             interpreter,
@@ -55,7 +54,6 @@ class RasaFileImporter(TrainingDataImporter):
             use_e2e,
             exclusion_percentage,
         )
-        return StoryGraph(story_steps)
 
     async def get_nlu_data(self, language: Optional[Text] = "en") -> TrainingData:
         return utils.training_data_from_paths(self._nlu_files, language)
